split exit list into separate component, camelcase json keys
