+++ out.log
@@ -1,0 +1,1 @@
+ash: /tmp/rita: not found
